application finie ?

+++ app/function/image.py
@@ -27,8 +27,9 @@ def update_image(destination, image, fk_id, db):
     table, fk_row_name, short_folder = ("Image_channel","id_channel_fk", SHORT_FOLDER_Y) if destination == "y" else (("Image_user","id_user_fk", SHORT_FOLDER_USER) if destination == "user" else ("Image_post","id_post_fk", SHORT_FOLDER_POST))
     old_image = db.execute(f"SELECT id_image, location FROM {table} WHERE {fk_row_name} = ?", (fk_id,)).fetchone()
 
-    if "default.png" not in old_image['location'] and image:
-        os.remove(os.path.join(WORKING_DIR + old_image['location']))
+    if old_image != None:
+        if "default.png" not in old_image['location'] and image:
+            os.remove(os.path.join(WORKING_DIR + old_image['location']))
     print("lol")
     try:
         if image and image.filename != '':
